+++ cmd/kube-proxy/app/server_others.go
@@ -47,6 +47,9 @@ import (
 	"k8s.io/component-base/configz"
 	"k8s.io/component-base/metrics"
 	utilsysctl "k8s.io/component-helpers/node/util/sysctl"
+	utilipset "k8s.io/kube-proxy/pkg/util/ipset"
+	utiliptables "k8s.io/kube-proxy/pkg/util/iptables"
+	utilipvs "k8s.io/kube-proxy/pkg/util/ipvs"
 	"k8s.io/kubernetes/pkg/proxy"
 	proxyconfigapi "k8s.io/kubernetes/pkg/proxy/apis/config"
 	"k8s.io/kubernetes/pkg/proxy/apis/config/scheme"
@@ -55,9 +58,6 @@ import (
 	"k8s.io/kubernetes/pkg/proxy/ipvs"
 	proxymetrics "k8s.io/kubernetes/pkg/proxy/metrics"
 	proxyutiliptables "k8s.io/kubernetes/pkg/proxy/util/iptables"
-	utilipset "k8s.io/kubernetes/pkg/util/ipset"
-	utiliptables "k8s.io/kubernetes/pkg/util/iptables"
-	utilipvs "k8s.io/kubernetes/pkg/util/ipvs"
 	utilnode "k8s.io/kubernetes/pkg/util/node"
 	"k8s.io/utils/exec"
 	netutils "k8s.io/utils/net"
