+++ cmd/kube-proxy/app/server_others_test.go
@@ -38,10 +38,10 @@ import (
 	clientsetfake "k8s.io/client-go/kubernetes/fake"
 	clientgotesting "k8s.io/client-go/testing"
 
+	utiliptables "k8s.io/kube-proxy/pkg/util/iptables"
+	utiliptablestest "k8s.io/kube-proxy/pkg/util/iptables/testing"
 	proxyconfigapi "k8s.io/kubernetes/pkg/proxy/apis/config"
 	proxyutiliptables "k8s.io/kubernetes/pkg/proxy/util/iptables"
-	utiliptables "k8s.io/kubernetes/pkg/util/iptables"
-	utiliptablestest "k8s.io/kubernetes/pkg/util/iptables/testing"
 )
 
 func Test_getDetectLocalMode(t *testing.T) {
