+++ pkg/proxy/ipvs/ipset_test.go
@@ -22,8 +22,8 @@ package ipvs
 import (
 	"testing"
 
-	utilipset "k8s.io/kubernetes/pkg/util/ipset"
-	fakeipset "k8s.io/kubernetes/pkg/util/ipset/testing"
+	utilipset "k8s.io/kube-proxy/pkg/util/ipset"
+	fakeipset "k8s.io/kube-proxy/pkg/util/ipset/testing"
 )
 
 func TestCheckIPSetVersion(t *testing.T) {
