+++ staging/publishing/import-restrictions.yaml
@@ -197,9 +197,11 @@
 
 - baseImportPath: "./vendor/k8s.io/kube-proxy/"
   allowedImports:
+  - k8s.io/api
   - k8s.io/apimachinery
   - k8s.io/component-base
   - k8s.io/klog
+  - k8s.io/kube-proxy
   - k8s.io/utils
 
 - baseImportPath: "./vendor/k8s.io/kube-scheduler/"
